implemented feature of  deletion and post details

+++ component/Post.tsx
@@ -1,5 +1,5 @@
 import { Box, Button, IconButton, Stack, Typography } from "@mui/material";
-import React from "react";
+import { useState } from "react";
 import { PostType } from "@/types/post";
 // import Image from "next/image";
 // import style from "@/styles/Post.module.css";
@@ -7,19 +7,26 @@ import DeleteIcon from "@mui/icons-material/Delete";
 import { useAtom } from "jotai";
 import { postAtom } from "@/pages/posts";
 import { deleteObject, ref, getStorage } from "firebase/storage";
+import { useSession } from "next-auth/react";
+import PostView from "./PostView";
 
 const Post = ({ post }: { post: PostType }) => {
   const { postId } = post;
   const [postList, setPostList] = useAtom(postAtom);
+  const { data, status } = useSession();
 
   const handleClick = async () => {
+    if (post.userEmail !== data?.user?.email) {
+      alert("Sorry ! You have not permisssion to delete Others post !");
+      return;
+    }
     try {
       const response = await fetch(`/api/posts/${postId}`, {
         method: "DELETE",
       });
       setPostList(postList.filter((post) => post.postId !== postId));
       const res = await response.json();
-      console.log("res: ", res);
+      // console.log("res: ", res);
       const storage = getStorage();
       const deleteRef = ref(storage, post.img);
       await deleteObject(deleteRef);
@@ -44,12 +51,35 @@ const Post = ({ post }: { post: PostType }) => {
         },
         "&:hover button": {
           visibility: "visible",
+          zIndex: "10000",
         },
         overflow: "hidden",
         position: "relative",
         borderRadius: "10px",
       }}
     >
+      <IconButton
+        aria-label="delete"
+        sx={{
+          color: "white",
+          border: "1px solid white",
+          width: "fit-content",
+          visibility: "hidden",
+          alignSelf: "end",
+          position: "absolute",
+          top: "20px",
+          right: "25px",
+          zIndex: "10000",
+          "&:hover": {
+            border: "1px solid red",
+            color: "red",
+          },
+        }}
+        size="small"
+        onClick={handleClick}
+      >
+        <DeleteIcon />
+      </IconButton>
       <Stack
         sx={{
           width: "100%",
@@ -69,27 +99,6 @@ const Post = ({ post }: { post: PostType }) => {
       >
         {/* <Image src={post.img} width={400} height={300} alt={`${post.title}-pic`} className={style.image}/> */}
 
-        <IconButton
-          aria-label="delete"
-          sx={{
-            color: "white",
-            border: "1px solid white",
-            width: "fit-content",
-            visibility: "hidden",
-            alignSelf: "end",
-            position: "absolute",
-            top: "20px",
-            right: "25px",
-            "&:hover": {
-              border: "1px solid red",
-              color: "red",
-            },
-          }}
-          size="small"
-          onClick={handleClick}
-        >
-          <DeleteIcon />
-        </IconButton>
         <Stack
           py={2}
           gap={2}
@@ -99,16 +108,6 @@ const Post = ({ post }: { post: PostType }) => {
             left: "20px",
           }}
         >
-          {/* <Typography
-            sx={{
-              textTransform: "capitalize",
-              fontSize: "1.5rem",
-              color: "white",
-              textShadow: "4px 4px 5px black",
-            }}
-          >
-            {post.title}
-          </Typography> */}
           <Stack
             direction="row"
             gap={2}
@@ -138,6 +137,7 @@ const Post = ({ post }: { post: PostType }) => {
           </Stack>
         </Stack>
       </Stack>
+      <PostView post={post} />
     </Box>
   );
 };
